III-1307: Rename PlaceProjectedToCdbXmlEventFactoryTest and adapt tests
